pyproject.toml
[tool.poetry]
name = "cc"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^3.2.0"
black = "^23.10.0"
isort = "^5.12.0"
pdbpp = "^0.10.3"
pytest-django = "^4.5.2"
bpython = "^0.24"
ruff = "^0.1.0"
python-dateutil = "^2.8.2"


[tool.poetry.group.dev.dependencies]
faker = "^19.11.0"
pytest-django = "^4.5.2"
pytest-mock = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
select = ["E", "F"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cctutorial.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
